Added GUI data saved to JSON
Now when the user clicks "save" on the main computer GUI, it's not only saving the logic for the configuration the user created, but also the GUI data, so it can be loaded again the next time the user opens the editor.

(For now, it only saves the data, it doesn't load it on startup. #TODO)